mivideo/requirements.txt
google-cloud-bigquery
pandas
